# dyno: Use llvm::SmallPtrSet instead of std::set for scope lookup