LLVM's SmallPtrSet is exactly what it says on the can: a set
optimzied for having few elements, which are pointers. Replacing
the more general `std::set` with LLVM's version seems to have a
positive impact on performance.

I've done some performance testing locally, and here are the
results in a release build chapel, while trying to scope resolve
_everything_ (including the internal modules, so not _just_
using `--dyno`).

```
old commit: 31419dc684d683dce821a54be1d34591ac2c6ac8
new commit: 74382d51ca446c641802010b3811a4fcba49b795
Benchmark 1: chapel-perf-old/bin/darwin-arm64/chpl perftest.chpl -o helpme --dyno
  Time (mean ± σ):     426.5 ms ±   2.2 ms    [User: 368.4 ms, System: 44.1 ms]
  Range (min … max):   422.6 ms … 430.4 ms    10 runs
 
Benchmark 2: chapel-perf-new/bin/darwin-arm64/chpl perftest.chpl -o helpme --dyno
  Time (mean ± σ):     374.3 ms ±   1.7 ms    [User: 316.1 ms, System: 44.8 ms]
  Range (min … max):   371.9 ms … 377.1 ms    10 runs
 
Summary
  '/chapel-perf-new/bin/darwin-arm64/chpl perftest.chpl -o helpme --dyno' ran
    1.14 ± 0.01 times faster than 'chapel-perf-old/bin/darwin-arm64/chpl perftest.chpl -o helpme --dyno'
```

So we're looking at roughly a 15% boost. This is only a local test, and until I can get a clearer picture of performance in general, I'm leaving this as a WIP PR.

__Important caveat to the performance measurements__: All other issues aside, note that Dyno scope resolving cannot handle all internal modules; the two executions that are timed are not representative of a full compiler execution, because the dyno scope resolver fails at some point in the execution.